\name{ADGofTest-package}
\alias{ADGofTest-package}
\alias{ADGofTest}
\docType{package}
\title{
    Implementation of the Anderson-Darling goodness of fit test.
}
\description{
    Implementation of the Anderson-Darling goodness of fit test.
}
\details{
\tabular{ll}{
Package: \tab ADGofTest\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-06-26\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Carlos J. Gil Bellosta

Maintainer: Carlos J. Gil Bellosta  <cjgb@datanalytics.com>
}
\references{
    G. and J. Marsaglia, "Evaluating the Anderson-Darling Distribution", Journal of Statistical Software, 2004
}
\keyword{ package }
